Move monitor code into a few separate modules & remove mocking
- Move most monitor code into BaseMonitor class.
- Put Windows code in Win32Monitor sub-class & confirm it works.
- Put X11 boilerplate code in place (not implemented yet).
- Adjust a few modules to import the monitor list and class for
  the current platform.
- Remove monitor-related mocking.
